Updates to harvest from Koha
Also disables / removes the unified discovery features. Koha grabs based on 999$c instead of 001, so the mapping to MODS is update so the 999$c will be indexed. Harvested records that do not exist will not be added; existing records WILL be updated. So primarily this is used to update the MARC is Islandora after the MARC in Koha is updated.